feat(python): split type stubs and generate agent-token classes
Move the hand-written API stubs into `_core.pyi` for the compiled
extension module, leaving `__init__.pyi` as a thin re-export list, and
have `generate_agent_tokens.py --update-stub` emit the agent-token
classes and `__all__` there so the package's shipped `py.typed` stubs
stay pinned to the same source of truth as the docs and Rust module.
Add a CI step that runs `stubtest` against the built module to catch
drift a diff cannot.
