+++ docs/special_tokens.md
@@ -1278,7 +1278,7 @@ One class per bundled vocabulary that carries agent tokens — every one except 
 
 Each also exposes its vocabulary's **native** markers, so one class is the whole special-token surface for that vocabulary rather than just splintr's half. Where a native name would collide with an agent token of the same meaning it takes a `_NATIVE` suffix — DeepSeek's `<think>` is `THINK_NATIVE` (128798), splintr's `<|think|>` is `THINK` (128905). Where the vocabulary defines the agent token _itself_, there is no suffix and no duplicate: `QWEN3_AGENT_TOKENS.IM_START` is Qwen's own 151644.
 
-These classes are generated by `scripts/generate_agent_tokens.py` into `src/python/agent_tokens_generated.rs`, and pinned against the tokenizers they name by `python/tests/test_agent_token_constants.py` — so a vocabulary added without regenerating fails a test rather than going missing quietly.
+These classes are generated by `scripts/generate_agent_tokens.py` into `src/python/agent_tokens_generated.rs`, and pinned against the tokenizers they name by `python/tests/test_agent_token_constants.py` — so a vocabulary added without regenerating fails a test rather than going missing quietly. The same script writes their type stubs (`--update-stub`, into `python/splintr/_core.pyi`); the package ships `py.typed`, so a stale stub makes a type checker reject a constant that exists, and CI fails on the drift.
 
 ### CL100K_AGENT_TOKENS
