+++ python/splintr/__init__.pyi
@@ -1,1495 +1,87 @@
-from typing import ClassVar, Final, NoReturn
+"""Type stubs for the `splintr` package.
+
+The API itself is described in `_core.pyi`, the compiled extension module this
+re-exports from. `__init__.py` re-exports the same names at runtime.
+"""
+
+from typing import Final
+
+from ._core import (
+    Tokenizer as Tokenizer,
+    SentencePieceTokenizer as SentencePieceTokenizer,
+    SpmTokenizer as SpmTokenizer,
+    WordPieceTokenizer as WordPieceTokenizer,
+    AnyTokenizer as AnyTokenizer,
+    from_json as from_json,
+    from_json_bytes as from_json_bytes,
+    base_vocab_size as base_vocab_size,
+    StreamingDecoder as StreamingDecoder,
+    CL100K_BASE_PATTERN as CL100K_BASE_PATTERN,
+    KIMI_PATTERN as KIMI_PATTERN,
+    QWEN2_PATTERN as QWEN2_PATTERN,
+    O200K_BASE_PATTERN as O200K_BASE_PATTERN,
+    LLAMA3_PATTERN as LLAMA3_PATTERN,
+    MISTRAL_V3_PATTERN as MISTRAL_V3_PATTERN,
+    GPT2_PATTERN as GPT2_PATTERN,
+    CL100K_AGENT_TOKENS as CL100K_AGENT_TOKENS,
+    O200K_AGENT_TOKENS as O200K_AGENT_TOKENS,
+    LLAMA3_AGENT_TOKENS as LLAMA3_AGENT_TOKENS,
+    DEEPSEEK_V3_AGENT_TOKENS as DEEPSEEK_V3_AGENT_TOKENS,
+    QWEN3_AGENT_TOKENS as QWEN3_AGENT_TOKENS,
+    GLM4_AGENT_TOKENS as GLM4_AGENT_TOKENS,
+    GPT_OSS_AGENT_TOKENS as GPT_OSS_AGENT_TOKENS,
+    KIMI_K2_AGENT_TOKENS as KIMI_K2_AGENT_TOKENS,
+    KIMI_K3_AGENT_TOKENS as KIMI_K3_AGENT_TOKENS,
+    MISTRAL_V1_AGENT_TOKENS as MISTRAL_V1_AGENT_TOKENS,
+    MISTRAL_V2_AGENT_TOKENS as MISTRAL_V2_AGENT_TOKENS,
+    MISTRAL_V3_AGENT_TOKENS as MISTRAL_V3_AGENT_TOKENS,
+    PHI4_AGENT_TOKENS as PHI4_AGENT_TOKENS,
+    OLMO2_AGENT_TOKENS as OLMO2_AGENT_TOKENS,
+    LLAMA2_AGENT_TOKENS as LLAMA2_AGENT_TOKENS,
+    CODELLAMA_AGENT_TOKENS as CODELLAMA_AGENT_TOKENS,
+    MODERNBERT_AGENT_TOKENS as MODERNBERT_AGENT_TOKENS,
+    GEMMA2_AGENT_TOKENS as GEMMA2_AGENT_TOKENS,
+    GEMMA3_AGENT_TOKENS as GEMMA3_AGENT_TOKENS,
+    GEMMA4_AGENT_TOKENS as GEMMA4_AGENT_TOKENS,
+)
 
 __version__: Final[str]
 
-CL100K_BASE_PATTERN: Final[str]
-KIMI_PATTERN: Final[str]
-QWEN2_PATTERN: Final[str]
-O200K_BASE_PATTERN: Final[str]
-LLAMA3_PATTERN: Final[str]
-MISTRAL_V3_PATTERN: Final[str]
-GPT2_PATTERN: Final[str]
-
-
-class StreamingDecoder:
-    def __new__(cls) -> NoReturn: ...
-    def add_token(self, token_id: int) -> str | None: ...
-    def add_tokens(self, token_ids: list[int]) -> str | None: ...
-    def flush(self) -> str: ...
-    def reset(self) -> None: ...
-    @property
-    def has_pending(self) -> bool: ...
-    @property
-    def pending_bytes(self) -> int: ...
-
-
-class Tokenizer:
-    def __init__(self, vocab_path: str, pattern: str, special_tokens: dict[str, int] | None = None) -> None: ...
-    @staticmethod
-    def from_pretrained(name: str) -> AnyTokenizer: ...
-    @staticmethod
-    def from_bytes(vocab_data: bytes, pattern: str, special_tokens: dict[str, int] | None = None) -> Tokenizer: ...
-    def pcre2(self, use_pcre2: bool = True) -> Tokenizer: ...
-    def jit(self, use_jit: bool = True) -> Tokenizer: ...
-    def encode(self, text: str) -> list[int]: ...
-    def encode_raw(self, text: str) -> list[int]: ...
-    def encode_rayon(self, text: str) -> list[int]: ...
-    def encode_with_special(self, text: str) -> list[int]: ...
-    def encode_ordinary(self, text: str) -> list[int]: ...
-    def encode_allowed_special(self, text: str, allowed_special: list[str]) -> list[int]: ...
-    def decode(self, tokens: list[int]) -> str: ...
-    def decode_with_special(self, tokens: list[int]) -> str: ...
-    def decode_bytes(self, tokens: list[int]) -> bytes: ...
-    def decode_lossy(self, tokens: list[int]) -> str: ...
-    def decode_token_bytes(self, id: int) -> bytes: ...
-    def decode_token(self, id: int) -> str: ...
-    def encode_batch(self, texts: list[str]) -> list[list[int]]: ...
-    def encode_flat(self, text: str) -> bytes: ...
-    def encode_batch_flat(self, texts: list[str]) -> tuple[bytes, bytes]: ...
-    def encode_batch_with_special(self, texts: list[str]) -> list[list[int]]: ...
-    def decode_batch(self, token_lists: list[list[int]]) -> list[str]: ...
-    def decode_batch_lossy(self, token_lists: list[list[int]]) -> list[str]: ...
-    @property
-    def vocab_size(self) -> int: ...
-    def streaming_decoder(self) -> StreamingDecoder: ...
-    def streaming_decoder_with_special(self) -> StreamingDecoder: ...
-    def clear_cache(self) -> None: ...
-    @property
-    def cache_len(self) -> int: ...
-
-
-class SentencePieceTokenizer:
-    def __init__(self, tokens: list[str], scores: list[float], eos_token_id: int, bos_token_id: int | None = None) -> None: ...
-    def encode(self, text: str) -> list[int]: ...
-    def encode_raw(self, text: str) -> list[int]: ...
-    def encode_with_special(self, text: str) -> list[int]: ...
-    def encode_ordinary(self, text: str) -> list[int]: ...
-    def encode_allowed_special(self, text: str, allowed_special: list[str]) -> list[int]: ...
-    def encode_batch(self, texts: list[str]) -> list[list[int]]: ...
-    def decode(self, ids: list[int]) -> str: ...
-    def decode_with_special(self, ids: list[int]) -> str: ...
-    def decode_lossy(self, ids: list[int]) -> str: ...
-    def decode_token_bytes(self, id: int) -> bytes: ...
-    def decode_token(self, id: int) -> str: ...
-    @property
-    def vocab_size(self) -> int: ...
-    def is_eos(self, token_id: int) -> bool: ...
-    @property
-    def eos_token_id(self) -> int: ...
-    @property
-    def bos_token_id(self) -> int | None: ...
-    def streaming_decoder(self) -> StreamingDecoder: ...
-    def streaming_decoder_with_special(self) -> StreamingDecoder: ...
-
-
-class SpmTokenizer:
-    def __init__(self, tokens: list[str], scores: list[float], bos_token_id: int | None = None, eos_token_id: int | None = None) -> None: ...
-    def encode(self, text: str) -> list[int]: ...
-    def encode_raw(self, text: str) -> list[int]: ...
-    def encode_with_special(self, text: str) -> list[int]: ...
-    def encode_batch(self, texts: list[str]) -> list[list[int]]: ...
-    def encode_ordinary(self, text: str) -> list[int]: ...
-    def encode_allowed_special(self, text: str, allowed_special: list[str]) -> list[int]: ...
-    def decode(self, ids: list[int]) -> str: ...
-    def decode_with_special(self, ids: list[int]) -> str: ...
-    def decode_token_bytes(self, id: int) -> bytes: ...
-    def decode_token(self, id: int) -> str: ...
-    @property
-    def vocab_size(self) -> int: ...
-    @property
-    def eos_token_id(self) -> int | None: ...
-    @property
-    def bos_token_id(self) -> int | None: ...
-    def streaming_decoder(self) -> StreamingDecoder: ...
-    def streaming_decoder_with_special(self) -> StreamingDecoder: ...
-
-
-class WordPieceTokenizer:
-    def __init__(self, vocab: list[str], unk_token_id: int, max_word_len: int = 100, do_lower_case: bool = False, strip_accents: bool | None = None) -> None: ...
-    def encode(self, text: str) -> list[int]: ...
-    def encode_raw(self, text: str) -> list[int]: ...
-    def encode_with_special(self, text: str) -> list[int]: ...
-    def encode_ordinary(self, text: str) -> list[int]: ...
-    def encode_allowed_special(self, text: str, allowed_special: list[str]) -> list[int]: ...
-    def encode_batch(self, texts: list[str]) -> list[list[int]]: ...
-    def decode(self, ids: list[int]) -> str: ...
-    def decode_with_special(self, ids: list[int]) -> str: ...
-    def decode_token_bytes(self, id: int) -> bytes: ...
-    def decode_token(self, id: int) -> str: ...
-    def vocab_size(self) -> int: ...
-    @property
-    def unk_token_id(self) -> int: ...
-    @property
-    def cls_token_id(self) -> int | None: ...
-    @property
-    def sep_token_id(self) -> int | None: ...
-    @property
-    def pad_token_id(self) -> int | None: ...
-    def streaming_decoder(self) -> StreamingDecoder: ...
-    def streaming_decoder_with_special(self) -> StreamingDecoder: ...
-
-
-class AnyTokenizer:
-    def __new__(cls) -> NoReturn: ...
-    def encode(self, text: str) -> list[int]: ...
-    def encode_raw(self, text: str) -> list[int]: ...
-    def encode_with_special(self, text: str) -> list[int]: ...
-    def encode_ordinary(self, text: str) -> list[int]: ...
-    def encode_allowed_special(self, text: str, allowed_special: list[str]) -> list[int]: ...
-    def encode_batch(self, texts: list[str]) -> list[list[int]]: ...
-    def encode_flat(self, text: str) -> bytes: ...
-    def encode_batch_flat(self, texts: list[str]) -> tuple[bytes, bytes]: ...
-    def encode_batch_with_special(self, texts: list[str]) -> list[list[int]]: ...
-    def encode_rayon(self, text: str) -> list[int]: ...
-    def decode(self, ids: list[int]) -> str: ...
-    def decode_with_special(self, ids: list[int]) -> str: ...
-    def decode_batch(self, token_lists: list[list[int]]) -> list[str]: ...
-    def decode_bytes(self, tokens: list[int]) -> bytes: ...
-    def decode_lossy(self, tokens: list[int]) -> str: ...
-    def decode_batch_lossy(self, token_lists: list[list[int]]) -> list[str]: ...
-    def decode_token_bytes(self, id: int) -> bytes: ...
-    def decode_token(self, id: int) -> str: ...
-    def pcre2(self, use_pcre2: bool = True) -> AnyTokenizer: ...
-    def jit(self, use_jit: bool = True) -> AnyTokenizer: ...
-    def streaming_decoder(self) -> StreamingDecoder: ...
-    def streaming_decoder_with_special(self) -> StreamingDecoder: ...
-    def clear_cache(self) -> None: ...
-    @property
-    def cache_len(self) -> int: ...
-    @property
-    def vocab_size(self) -> int: ...
-    def is_eos(self, token_id: int) -> bool: ...
-    @property
-    def eos_token_id(self) -> int | None: ...
-    def special_token_id(self, name: str) -> int | None: ...
-    def special_tokens(self) -> dict[str, int]: ...
-    @property
-    def family(self) -> str: ...
-
-
-def from_json(path: str) -> AnyTokenizer: ...
-def from_json_bytes(data: bytes) -> AnyTokenizer: ...
-def base_vocab_size(name: str) -> int: ...
-
-
-class CL100K_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class O200K_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class LLAMA3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    BEGIN_OF_TEXT: ClassVar[int]
-    END_OF_TEXT: ClassVar[int]
-    FINETUNE_RIGHT_PAD_ID: ClassVar[int]
-    STEP_ID: ClassVar[int]
-    START_HEADER_ID: ClassVar[int]
-    END_HEADER_ID: ClassVar[int]
-    EOM_ID: ClassVar[int]
-    EOT_ID: ClassVar[int]
-    PYTHON_TAG: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class DEEPSEEK_V3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    BEGIN_OF_SENTENCE: ClassVar[int]
-    END_OF_SENTENCE: ClassVar[int]
-    PAD_NATIVE: ClassVar[int]
-    THINK_NATIVE: ClassVar[int]
-    THINK_END_NATIVE: ClassVar[int]
-    FIM_HOLE: ClassVar[int]
-    FIM_BEGIN: ClassVar[int]
-    FIM_END: ClassVar[int]
-    USER_NATIVE: ClassVar[int]
-    ASSISTANT_NATIVE: ClassVar[int]
-    EOT: ClassVar[int]
-    TOOL_CALLS_BEGIN: ClassVar[int]
-    TOOL_CALLS_END: ClassVar[int]
-    TOOL_CALL_BEGIN: ClassVar[int]
-    TOOL_CALL_END: ClassVar[int]
-    TOOL_OUTPUTS_BEGIN: ClassVar[int]
-    TOOL_OUTPUTS_END: ClassVar[int]
-    TOOL_OUTPUT_BEGIN: ClassVar[int]
-    TOOL_OUTPUT_END: ClassVar[int]
-    TOOL_SEP: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class MISTRAL_V1_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class MISTRAL_V2_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class MISTRAL_V3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    INST: ClassVar[int]
-    INST_END: ClassVar[int]
-    AVAILABLE_TOOLS: ClassVar[int]
-    AVAILABLE_TOOLS_END: ClassVar[int]
-    TOOL_RESULTS: ClassVar[int]
-    TOOL_RESULTS_END: ClassVar[int]
-    TOOL_CALLS: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class QWEN3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    ENDOFTEXT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    OBJECT_REF_START: ClassVar[int]
-    OBJECT_REF_END: ClassVar[int]
-    BOX_START: ClassVar[int]
-    BOX_END: ClassVar[int]
-    QUAD_START: ClassVar[int]
-    QUAD_END: ClassVar[int]
-    VISION_START: ClassVar[int]
-    VISION_END: ClassVar[int]
-    VISION_PAD: ClassVar[int]
-    IMAGE_PAD: ClassVar[int]
-    VIDEO_PAD: ClassVar[int]
-    TOOL_CALL: ClassVar[int]
-    TOOL_CALL_END: ClassVar[int]
-    FIM_PREFIX: ClassVar[int]
-    FIM_MIDDLE: ClassVar[int]
-    FIM_SUFFIX: ClassVar[int]
-    FIM_PAD: ClassVar[int]
-    REPO_NAME: ClassVar[int]
-    FILE_SEP: ClassVar[int]
-    TOOL_RESPONSE: ClassVar[int]
-    TOOL_RESPONSE_END: ClassVar[int]
-    THINK_NATIVE: ClassVar[int]
-    THINK_END_NATIVE: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class GLM4_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    ENDOFTEXT: ClassVar[int]
-    MASK: ClassVar[int]
-    GMASK: ClassVar[int]
-    SMASK: ClassVar[int]
-    SOP: ClassVar[int]
-    EOP: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    OBSERVATION: ClassVar[int]
-    BEGIN_OF_IMAGE: ClassVar[int]
-    END_OF_IMAGE: ClassVar[int]
-    BEGIN_OF_VIDEO: ClassVar[int]
-    END_OF_VIDEO: ClassVar[int]
-    BEGIN_OF_AUDIO: ClassVar[int]
-    END_OF_AUDIO: ClassVar[int]
-    BEGIN_OF_TRANSCRIPTION: ClassVar[int]
-    END_OF_TRANSCRIPTION: ClassVar[int]
-    CODE_PREFIX: ClassVar[int]
-    CODE_MIDDLE: ClassVar[int]
-    CODE_SUFFIX: ClassVar[int]
-    THINK_NATIVE: ClassVar[int]
-    THINK_END_NATIVE: ClassVar[int]
-    TOOL_CALL: ClassVar[int]
-    TOOL_CALL_END: ClassVar[int]
-    TOOL_RESPONSE: ClassVar[int]
-    TOOL_RESPONSE_END: ClassVar[int]
-    ARG_KEY: ClassVar[int]
-    ARG_KEY_END: ClassVar[int]
-    ARG_VALUE: ClassVar[int]
-    ARG_VALUE_END: ClassVar[int]
-    NOTHINK: ClassVar[int]
-    BEGIN_OF_BOX: ClassVar[int]
-    END_OF_BOX: ClassVar[int]
-    IMAGE: ClassVar[int]
-    VIDEO: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class GPT_OSS_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    STARTOFTEXT: ClassVar[int]
-    ENDOFTEXT: ClassVar[int]
-    RETURN: ClassVar[int]
-    CONSTRAIN: ClassVar[int]
-    CHANNEL: ClassVar[int]
-    START: ClassVar[int]
-    END: ClassVar[int]
-    MESSAGE: ClassVar[int]
-    CALL: ClassVar[int]
-    ENDOFPROMPT: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class KIMI_K2_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    BOS: ClassVar[int]
-    EOS: ClassVar[int]
-    IM_END: ClassVar[int]
-    IM_USER: ClassVar[int]
-    IM_ASSISTANT: ClassVar[int]
-    START_HEADER_ID: ClassVar[int]
-    END_HEADER_ID: ClassVar[int]
-    EOT: ClassVar[int]
-    IM_SYSTEM: ClassVar[int]
-    TOOL_CALLS_SECTION_BEGIN: ClassVar[int]
-    TOOL_CALLS_SECTION_END: ClassVar[int]
-    TOOL_CALL_BEGIN: ClassVar[int]
-    TOOL_CALL_ARGUMENT_BEGIN: ClassVar[int]
-    TOOL_CALL_END: ClassVar[int]
-    IM_MIDDLE: ClassVar[int]
-    MEDIA_BEGIN: ClassVar[int]
-    MEDIA_CONTENT: ClassVar[int]
-    MEDIA_END: ClassVar[int]
-    MEDIA_PAD: ClassVar[int]
-    THINK_NATIVE: ClassVar[int]
-    THINK_END_NATIVE: ClassVar[int]
-    UNK: ClassVar[int]
-    PAD_NATIVE: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class KIMI_K3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    BOS: ClassVar[int]
-    EOS: ClassVar[int]
-    END_OF_MSG: ClassVar[int]
-    OPEN: ClassVar[int]
-    CLOSE: ClassVar[int]
-    SEP: ClassVar[int]
-    START_HEADER_ID: ClassVar[int]
-    END_HEADER_ID: ClassVar[int]
-    EOT: ClassVar[int]
-    MEDIA_BEGIN: ClassVar[int]
-    MEDIA_CONTENT: ClassVar[int]
-    MEDIA_END: ClassVar[int]
-    MEDIA_PAD: ClassVar[int]
-    OSAGENT_MODE: ClassVar[int]
-    UNK: ClassVar[int]
-    PAD_NATIVE: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class PHI4_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    ENDOFTEXT: ClassVar[int]
-    FIM_PREFIX: ClassVar[int]
-    FIM_MIDDLE: ClassVar[int]
-    FIM_SUFFIX: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    IM_SEP: ClassVar[int]
-    ENDOFPROMPT: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class OLMO2_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    EXTRA_ID_0: ClassVar[int]
-    ENDOFTEXT: ClassVar[int]
-    FIM_PREFIX: ClassVar[int]
-    FIM_MIDDLE: ClassVar[int]
-    FIM_SUFFIX: ClassVar[int]
-    PHONE_NUMBER: ClassVar[int]
-    EMAIL_ADDRESS: ClassVar[int]
-    IP_ADDRESS: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    EXTRA_ID_1: ClassVar[int]
-    EXTRA_ID_2: ClassVar[int]
-    EXTRA_ID_3: ClassVar[int]
-    EXTRA_ID_4: ClassVar[int]
-    EXTRA_ID_5: ClassVar[int]
-    EXTRA_ID_6: ClassVar[int]
-    EXTRA_ID_7: ClassVar[int]
-    EXTRA_ID_8: ClassVar[int]
-    EXTRA_ID_9: ClassVar[int]
-    EXTRA_ID_10: ClassVar[int]
-    ENDOFPROMPT: ClassVar[int]
-    PAD: ClassVar[int]
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class LLAMA2_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class CODELLAMA_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class MODERNBERT_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class GEMMA2_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class GEMMA3_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
-
-
-class GEMMA4_AGENT_TOKENS:
-    def __new__(cls) -> NoReturn: ...
-    SYSTEM: ClassVar[int]
-    USER: ClassVar[int]
-    ASSISTANT: ClassVar[int]
-    IM_START: ClassVar[int]
-    IM_END: ClassVar[int]
-    THINK: ClassVar[int]
-    THINK_END: ClassVar[int]
-    PLAN: ClassVar[int]
-    PLAN_END: ClassVar[int]
-    STEP: ClassVar[int]
-    STEP_END: ClassVar[int]
-    ACT: ClassVar[int]
-    ACT_END: ClassVar[int]
-    OBSERVE: ClassVar[int]
-    OBSERVE_END: ClassVar[int]
-    FUNCTION: ClassVar[int]
-    FUNCTION_END: ClassVar[int]
-    RESULT: ClassVar[int]
-    RESULT_END: ClassVar[int]
-    ERROR: ClassVar[int]
-    ERROR_END: ClassVar[int]
-    CODE: ClassVar[int]
-    CODE_END: ClassVar[int]
-    OUTPUT: ClassVar[int]
-    OUTPUT_END: ClassVar[int]
-    LANG: ClassVar[int]
-    LANG_END: ClassVar[int]
-    CONTEXT: ClassVar[int]
-    CONTEXT_END: ClassVar[int]
-    QUOTE: ClassVar[int]
-    QUOTE_END: ClassVar[int]
-    CITE: ClassVar[int]
-    CITE_END: ClassVar[int]
-    SOURCE: ClassVar[int]
-    SOURCE_END: ClassVar[int]
-    MEMORY: ClassVar[int]
-    MEMORY_END: ClassVar[int]
-    RECALL: ClassVar[int]
-    RECALL_END: ClassVar[int]
-    PAD: ClassVar[int]
-    STOP: ClassVar[int]
-    SEP: ClassVar[int]
-    IMAGE: ClassVar[int]
-    IMAGE_END: ClassVar[int]
-    AUDIO: ClassVar[int]
-    AUDIO_END: ClassVar[int]
-    VIDEO: ClassVar[int]
-    VIDEO_END: ClassVar[int]
-    TITLE: ClassVar[int]
-    TITLE_END: ClassVar[int]
-    SECTION: ClassVar[int]
-    SECTION_END: ClassVar[int]
-    SUMMARY: ClassVar[int]
-    SUMMARY_END: ClassVar[int]
+__all__ = [
+    "Tokenizer",
+    "SentencePieceTokenizer",
+    "SpmTokenizer",
+    "WordPieceTokenizer",
+    "AnyTokenizer",
+    "from_json",
+    "from_json_bytes",
+    "base_vocab_size",
+    "StreamingDecoder",
+    "CL100K_BASE_PATTERN",
+    "KIMI_PATTERN",
+    "QWEN2_PATTERN",
+    "O200K_BASE_PATTERN",
+    "LLAMA3_PATTERN",
+    "MISTRAL_V3_PATTERN",
+    "GPT2_PATTERN",
+    "CL100K_AGENT_TOKENS",
+    "O200K_AGENT_TOKENS",
+    "LLAMA3_AGENT_TOKENS",
+    "DEEPSEEK_V3_AGENT_TOKENS",
+    "QWEN3_AGENT_TOKENS",
+    "GLM4_AGENT_TOKENS",
+    "GPT_OSS_AGENT_TOKENS",
+    "KIMI_K2_AGENT_TOKENS",
+    "KIMI_K3_AGENT_TOKENS",
+    "MISTRAL_V1_AGENT_TOKENS",
+    "MISTRAL_V2_AGENT_TOKENS",
+    "MISTRAL_V3_AGENT_TOKENS",
+    "PHI4_AGENT_TOKENS",
+    "OLMO2_AGENT_TOKENS",
+    "LLAMA2_AGENT_TOKENS",
+    "CODELLAMA_AGENT_TOKENS",
+    "MODERNBERT_AGENT_TOKENS",
+    "GEMMA2_AGENT_TOKENS",
+    "GEMMA3_AGENT_TOKENS",
+    "GEMMA4_AGENT_TOKENS",
+]
